% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kinesisanalytics_package.R
\docType{package}
\name{paws.kinesisanalytics-package}
\alias{paws.kinesisanalytics}
\alias{paws.kinesisanalytics-package}
\title{paws.kinesisanalytics: Amazon Kinesis Analytics}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
