% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kinesisanalytics_operations.R
\name{update_application}
\alias{update_application}
\title{Updates an existing Amazon Kinesis Analytics application}
\usage{
update_application(ApplicationName, CurrentApplicationVersionId,
  ApplicationUpdate)
}
\arguments{
\item{ApplicationName}{Name of the Amazon Kinesis Analytics application to update.}

\item{CurrentApplicationVersionId}{The current application version ID. You can use the DescribeApplication operation to get this value.}

\item{ApplicationUpdate}{Describes application updates.}
}
\description{
Updates an existing Amazon Kinesis Analytics application. Using this API, you can update application code, input configuration, and output configuration.
}
\details{
Note that Amazon Kinesis Analytics updates the \code{CurrentApplicationVersionId} each time you update your application.

This operation requires permission for the \code{kinesisanalytics:UpdateApplication} action.
}
\examples{

}
