% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.kinesisanalytics_operations.R
\name{delete_application_input_processing_configuration}
\alias{delete_application_input_processing_configuration}
\title{Deletes an InputProcessingConfiguration from an input}
\usage{
delete_application_input_processing_configuration(ApplicationName,
  CurrentApplicationVersionId, InputId)
}
\arguments{
\item{ApplicationName}{The Kinesis Analytics application name.}

\item{CurrentApplicationVersionId}{The version ID of the Kinesis Analytics application.}

\item{InputId}{The ID of the input configuration from which to delete the input processing configuration. You can get a list of the input IDs for an application by using the DescribeApplication operation.}
}
\description{
Deletes an InputProcessingConfiguration from an input.
}
\examples{

}
